feat: implement basic merkle set and merkle tree structs

+++ core/src/utils/mod.rs
@@ -9,13 +9,13 @@ pub trait ToElements {
     fn to_elements(&self) -> Vec<BaseElement>;
 }
 
-impl<const N: usize> ToElements for [u128; N] {
+impl<const N: usize> ToElements for [u64; N] {
     fn to_elements(&self) -> Vec<BaseElement> {
         self.iter().map(|&v| BaseElement::new(v)).collect()
     }
 }
 
-impl ToElements for Vec<u128> {
+impl ToElements for Vec<u64> {
     fn to_elements(&self) -> Vec<BaseElement> {
         self.iter().map(|&v| BaseElement::new(v)).collect()
     }
